Split out AttributeResult and EnumAttributeResult to separate files.
