+++ searchlib/src/vespa/searchlib/expression/CMakeLists.txt
@@ -1,6 +1,9 @@
 # Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(searchlib_expression OBJECT
     SOURCES
+    attributenode.cpp
+    attributeresult.cpp
+    enumattributeresult.cpp
     perdocexpression.cpp
     expressiontree.cpp
     timestamp.cpp
@@ -15,7 +18,6 @@ vespa_add_library(searchlib_expression OBJECT
     catserializer.cpp
     strcatserializer.cpp
     documentfieldnode.cpp
-    attributenode.cpp
     zcurve.cpp
     debugwaitfunctionnode.cpp
     mathfunctionnode.cpp
