+++ searchlib/src/vespa/searchlib/expression/attributenode.h
@@ -3,40 +3,13 @@
 
 #include "functionnode.h"
 #include "resultvector.h"
-#include <vespa/searchcommon/attribute/iattributecontext.h>
+#include "attributeresult.h"
 #include <vespa/vespalib/objects/objectoperation.h>
 #include <vespa/vespalib/objects/objectpredicate.h>
 
+namespace search::attribute { class IAttributeContext; }
+
 namespace search::expression {
-
-class AttributeResult : public ResultNode
-{
-public:
-    typedef std::unique_ptr<AttributeResult> UP;
-    DECLARE_RESULTNODE(AttributeResult);
-    AttributeResult() : _attribute(NULL), _docId(0) { }
-    AttributeResult(const attribute::IAttributeVector * attribute, DocId docId) :
-        _attribute(attribute),
-        _docId(docId)
-    { }
-    void setDocId(DocId docId) { _docId = docId; }
-    const search::attribute::IAttributeVector *getAttribute() const { return _attribute; }
-    DocId getDocId() const { return _docId; }
-private:
-    int64_t onGetInteger(size_t index) const override { (void) index; return _attribute->getInt(_docId); }
-    double onGetFloat(size_t index)    const override { (void) index; return _attribute->getFloat(_docId); }
-    ConstBufferRef onGetString(size_t index, BufferRef buf) const override {
-        (void) index;
-        const char * t = _attribute->getString(_docId, buf.str(), buf.size());
-        return ConstBufferRef(t, strlen(t));
-    }
-    int64_t onGetEnum(size_t index) const override { (void) index; return (static_cast<int64_t>(_attribute->getEnum(_docId))); }
-    void set(const search::expression::ResultNode&) override { }
-    size_t hash() const override { return _docId; }
-
-    const search::attribute::IAttributeVector * _attribute;
-    DocId          _docId;
-};
 
 class AttributeNode : public FunctionNode
 {
@@ -146,11 +119,11 @@ private:
         mutable std::vector<search::attribute::IAttributeVector::WeightedEnum> _wVector;
     };
 
-    AttributeResult::UP         _scratchResult;
-    bool                        _hasMultiValue;
-    bool                        _useEnumOptimization;
-    std::unique_ptr<Handler>    _handler;
-    vespalib::string            _attributeName;
+    std::unique_ptr<AttributeResult> _scratchResult;
+    bool                             _hasMultiValue;
+    bool                             _useEnumOptimization;
+    std::unique_ptr<Handler>         _handler;
+    vespalib::string                 _attributeName;
 };
 
 }
