+++ searchlib/src/vespa/searchlib/expression/attributeresult.cpp
@@ -1,0 +1,9 @@
+// Copyright 2018 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#include "attributeresult.h"
+
+namespace search::expression {
+
+IMPLEMENT_RESULTNODE(AttributeResult, ResultNode);
+
+}
